remove browsable api auth
This was conflicting with caching of pages, as the logged in user was shown in the upper right.
To resolve either a vary_on_headers was needed, which would affect caching of json api, causing the cache to only be per user. or to do some custom logic to cache the actual json contents from the serializer

Cacheing the json from the serializer is the method that should be done in the future.